+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 4.0 (unreleased)
 ----------------
 
+- Implement new `IQueryIndex` interface.
+
 - Require at least Zope and ZCatalog 4.0.
 
 3.0 (2016-07-18)